remove MultipleDataSourcesRuntimeContext
