+++ sharding-jdbc/sharding-jdbc-core/src/main/java/org/apache/shardingsphere/shardingjdbc/jdbc/core/context/MasterSlaveRuntimeContext.java
@@ -35,7 +35,7 @@ import java.util.Properties;
  * Runtime context for master-slave.
  */
 @Getter
-public final class MasterSlaveRuntimeContext extends MultipleDataSourcesRuntimeContext<MasterSlaveRule> {
+public final class MasterSlaveRuntimeContext extends AbstractRuntimeContext<MasterSlaveRule> {
     
     private final CachedDatabaseMetaData cachedDatabaseMetaData;
