+++ sharding-jdbc/sharding-jdbc-core/src/main/java/org/apache/shardingsphere/shardingjdbc/jdbc/core/context/ShadowRuntimeContext.java
@@ -35,7 +35,7 @@ import java.util.Properties;
  * Runtime context for shadow.
  */
 @Getter
-public final class ShadowRuntimeContext extends MultipleDataSourcesRuntimeContext<ShadowRule> {
+public final class ShadowRuntimeContext extends AbstractRuntimeContext<ShadowRule> {
     
     private final DataSource actualDataSource;
